+++ guillotina_elasticsearch/tests/test_schema.py
@@ -8,9 +8,6 @@ from zope.interface import Interface
 import pytest
 
 
-pytestmark = [pytest.mark.asyncio]
-
-
 class IA(Interface):
     index_field("item", field_mapping={"type": "integer"})
     item = Int()